Move get_packit_commands_from_comment and CeleryTask class
They both are quite independent, so they don't need to be in the abstract.py.
This should avoid circular imports.

Fixes #1618
