+++ packit_service/utils.py
@@ -152,3 +152,23 @@ def get_timezone_aware_datetime(datetime_to_update: datetime) -> datetime:
         timezone-aware datetime
     """
     return datetime_to_update.replace(tzinfo=timezone.utc)
+
+
+def get_packit_commands_from_comment(
+    comment: str, packit_comment_command_prefix: str
+) -> List[str]:
+    comment_parts = comment.strip()
+
+    if not comment_parts:
+        logger.debug("Empty comment, nothing to do.")
+        return []
+
+    comment_lines = comment_parts.split("\n")
+
+    for line in filter(None, map(str.strip, comment_lines)):
+        (packit_mark, *packit_command) = line.split(maxsplit=3)
+        # packit_command[0] has the first cmd and [1] has the second, if needed.
+        if packit_mark == packit_comment_command_prefix and packit_command:
+            return packit_command
+
+    return []
