+++ packit_service/worker/handlers/abstract.py
@@ -7,30 +7,30 @@ This file defines generic job handler
 import enum
 import logging
 import shutil
-from celery import Task
 from collections import defaultdict
 from datetime import datetime
 from os import getenv
 from pathlib import Path
-from typing import Dict, Optional, Set, Type, List
+from typing import Dict, Optional, Set, Type
 
+from celery import Task
 from celery import signature
 from celery.canvas import Signature
 from ogr.abstract import GitProject
+
 from packit.api import PackitAPI
 from packit.config import JobConfig, JobType, PackageConfig
 from packit.constants import DATETIME_FORMAT
 from packit.local_project import LocalProject
-
 from packit_service.config import ServiceConfig
-from packit_service.constants import DEFAULT_RETRY_LIMIT
 from packit_service.models import (
     AbstractTriggerDbType,
 )
 from packit_service.sentry_integration import push_scope_to_sentry
+from packit_service.utils import dump_job_config, dump_package_config
+from packit_service.worker.celery_task import CeleryTask
 from packit_service.worker.events import Event, EventData
 from packit_service.worker.monitoring import Pushgateway
-from packit_service.utils import dump_job_config, dump_package_config
 from packit_service.worker.result import TaskResults
 
 logger = logging.getLogger(__name__)
@@ -189,88 +189,6 @@ def run_for_check_rerun(prefix: str):
         return kls
 
     return _add_to_mapping
-
-
-def get_packit_commands_from_comment(
-    comment: str, packit_comment_command_prefix: str
-) -> List[str]:
-    comment_parts = comment.strip()
-
-    if not comment_parts:
-        logger.debug("Empty comment, nothing to do.")
-        return []
-
-    comment_lines = comment_parts.split("\n")
-
-    for line in filter(None, map(str.strip, comment_lines)):
-        (packit_mark, *packit_command) = line.split(maxsplit=3)
-        # packit_command[0] has the first cmd and [1] has the second, if needed.
-        if packit_mark == packit_comment_command_prefix and packit_command:
-            return packit_command
-
-    return []
-
-
-class CeleryTask:
-    """
-    Class wrapping the Celery task object with methods related to retrying.
-    """
-
-    def __init__(self, task: Task):
-        self.task = task
-
-    @property
-    def retries(self):
-        return self.task.request.retries
-
-    def is_last_try(self) -> bool:
-        """
-        Returns True if the current celery task is run for the last try.
-        More info about retries can be found here:
-        https://docs.celeryq.dev/en/latest/userguide/tasks.html#retrying
-        """
-        return self.retries >= self.get_retry_limit()
-
-    def get_retry_limit(self) -> int:
-        """
-        Returns the limit of the celery task retries.
-        (Packit uses this env.var. in HandlerTaskWithRetry base class
-        to set `max_retries` in `retry_kwargs`.)
-        """
-        return int(getenv("CELERY_RETRY_LIMIT", DEFAULT_RETRY_LIMIT))
-
-    def retry(
-        self,
-        ex: Optional[Exception] = None,
-        delay: Optional[int] = None,
-        max_retries: Optional[int] = None,
-    ) -> None:
-        """
-        Retries the celery task.
-        Argument `throw` is set to False to not retry
-        the task also because of the `autoretry_for` mechanism.
-
-        More info about retries can be found here:
-        https://docs.celeryq.dev/en/latest/userguide/tasks.html#retrying
-
-        Args:
-            ex: Exception which caused the retry (will be logged).
-            delay: Number of seconds the task will wait before being run again.
-            max_retries: Maximum number of retries to use instead of the default within
-                HandlerTaskWithRetry.
-        """
-        retries = self.retries
-        delay = delay if delay is not None else 60 * 2**retries
-        logger.info(f"Will retry for the {retries + 1}. time in {delay}s.")
-        kargs = self.task.request.kwargs.copy()
-        self.task.retry(
-            exc=ex,
-            countdown=delay,
-            throw=False,
-            args=(),
-            kwargs=kargs,
-            max_retries=max_retries,
-        )
 
 
 class TaskName(str, enum.Enum):
