+++ packit_service/worker/helpers/testing_farm.py
@@ -32,17 +32,14 @@ from packit_service.models import (
 )
 from packit_service.sentry_integration import send_to_sentry
 from packit_service.service.urls import get_testing_farm_info_url
-from packit_service.utils import get_package_nvrs
-from packit_service.worker.handlers.abstract import (
-    CeleryTask,
-    get_packit_commands_from_comment,
-)
 from packit_service.worker.events import (
     EventData,
     PullRequestCommentGithubEvent,
     MergeRequestCommentGitlabEvent,
     PullRequestCommentPagureEvent,
 )
+from packit_service.utils import get_package_nvrs, get_packit_commands_from_comment
+from packit_service.worker.celery_task import CeleryTask
 from packit_service.worker.helpers.build import CoprBuildJobHelper
 from packit_service.worker.reporting import BaseCommitStatus
 from packit_service.worker.result import TaskResults
